[FLINK-36602][table] Upgrade Calcite version to 1.38.0

+++ flink-table/flink-table-planner/src/main/scala/org/apache/flink/table/planner/plan/utils/FlinkRexUtil.scala
@@ -348,7 +348,7 @@ object FlinkRexUtil {
               "null:" + l.getType
             } else {
               l.getTypeName match {
-                case SqlTypeName.DOUBLE => Util.toScientificNotation(value.asInstanceOf[BigDecimal])
+                case SqlTypeName.DOUBLE => Util.toScientificNotation(value.asInstanceOf[Double])
                 case SqlTypeName.BIGINT => s"${value.asInstanceOf[BigDecimal].longValue()}L"
                 case SqlTypeName.BINARY => s"X'${value.asInstanceOf[ByteString].toString(16)}'"
                 case SqlTypeName.VARCHAR | SqlTypeName.CHAR =>
